Fixes and polishes
