+++ .env.example
@@ -2,5 +2,5 @@ GOOGLE_GENERATIVE_AI_API_KEY=****
 OPENAI_API_KEY=****
 XAI_API_KEY=****
 ANTHROPIC_API_KEY=****
-POSTGRES_URL=postgres://postgres:postgres@localhost:5432/chatbot
+POSTGRES_URL=postgres://your_username:your_password@postgres:5432/your_database_name
 AUTH_SECRET="****" # Added by `npx auth secret`. Read more: https://cli.authjs.dev
