+++ docker/compose.yml
@@ -1,0 +1,37 @@
+services:
+  mcp-client-chatbot:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - '3000:3000'
+    environment:
+      - AUTH_TRUST_HOST=true
+      - NO_HTTPS=true
+    env_file:
+      - ../.env
+    dns:
+      - 8.8.8.8 # Google's public DNS server
+      - 8.8.4.4 # Google's public DNS server
+    networks:
+      - mcp_client_network
+    restart: unless-stopped
+
+  postgres:
+    image: postgres:17
+    environment:
+      POSTGRES_DB: your_database_name 
+      POSTGRES_USER: your_username    
+      POSTGRES_PASSWORD: your_password 
+    networks:
+      - mcp_client_network
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+    restart: unless-stopped
+
+volumes:
+  postgres_data:
+
+networks:
+  mcp_client_network:
+    driver: bridge
