+++ src/components/mcp-editor.tsx
@@ -16,7 +16,7 @@ import { useRouter } from "next/navigation";
 import { createDebounce, isNull, safeJSONParse } from "lib/utils";
 import { handleErrorWithToast } from "ui/shared-toast";
 import { mutate } from "swr";
-import { Loader, AlertTriangle } from "lucide-react";
+import { Loader } from "lucide-react";
 import {
   isMaybeMCPServerConfig,
   isMaybeSseConfig,
@@ -53,7 +53,6 @@ export default function MCPEditor({
   initialConfig,
   name: initialName,
 }: MCPEditorProps) {
-  const isVercelEnv = process.env.NEXT_PUBLIC_VERCEL === "1";
   const shouldInsert = useMemo(() => isNull(initialName), [initialName]);
   const [isLoading, setIsLoading] = useState(false);
   const [jsonError, setJsonError] = useState<string | null>(null);
@@ -101,20 +100,6 @@ export default function MCPEditor({
 
   // Validate
   const validateConfig = (jsonConfig: unknown): boolean => {
-    // Check if we're in Vercel environment and config contains a command
-    if (
-      process.env.NEXT_PUBLIC_VERCEL === "1" &&
-      typeof jsonConfig === "object" &&
-      jsonConfig !== null &&
-      "command" in jsonConfig
-    ) {
-      handleErrorWithToast(
-        new Error("STDIO is not supported in the Vercel environment"),
-        "mcp-editor-error",
-      );
-      return false;
-    }
-
     const result = isMaybeSseConfig(jsonConfig)
       ? MCPSseConfigZodSchema.safeParse(jsonConfig)
       : MCPStdioConfigZodSchema.safeParse(jsonConfig);
@@ -163,26 +148,13 @@ export default function MCPEditor({
     errorDebounce.clear();
     if (result.success) {
       const isDiff = !equal(result.value, config);
-
-      // Check if we're in Vercel environment and config contains a command
-      if (
-        process.env.NEXT_PUBLIC_VERCEL === "1" &&
-        typeof result.value === "object" &&
-        result.value !== null &&
-        "command" in result.value
-      ) {
-        setJsonError(
-          "Command execution is not allowed in the Vercel environment",
+      setConfig(result.value as MCPServerConfig);
+      setJsonError(null);
+      if (isDiff) {
+        convertDebounce(
+          () => setJsonString(JSON.stringify(result.value, null, 2)),
+          5000,
         );
-      } else {
-        setConfig(result.value as MCPServerConfig);
-        setJsonError(null);
-        if (isDiff) {
-          convertDebounce(
-            () => setJsonString(JSON.stringify(result.value, null, 2)),
-            5000,
-          );
-        }
       }
     } else if (data.trim() !== "") {
       errorDebounce(() => {
@@ -216,16 +188,6 @@ export default function MCPEditor({
       <div className="space-y-4">
         <div className="space-y-2">
           <Label htmlFor="config">Config</Label>
-          {isVercelEnv && (
-            <Alert className="mb-2">
-              <AlertTriangle className="h-4 w-4" />
-              <AlertTitle>Vercel Environment Detected</AlertTitle>
-              <AlertDescription>
-                Command execution configurations are not allowed in the Vercel
-                environment. Please use SSE configurations instead.
-              </AlertDescription>
-            </Alert>
-          )}
         </div>
 
         {/* Split view for config editor */}
